+++ src/components/home/Calendar.js
@@ -1,25 +1,13 @@
-import React from "react";
+import React, { useState } from "react";
+import Calendar from "react-calendar";
+import "react-calendar/dist/Calendar.css";
 
-function Calendar() {
+export default function CalendarArea() {
+  const [value, onChange] = useState(new Date());
+
   return (
-    <div className="md:pr-12 mb-16">
-      <p className="text-justify px-4 md:px-0">
-        Lorem Ipsum is simply dummy text of the printing and typesetting
-        industry. Lorem Ipsum has been the industry's standard dummy text ever
-        since the 1500s, when an unknown printer took a galley of type and
-        scrambled it to make a type specimen book. It has survived not only five
-        centuries, but also the leap into electronic typesetting, remaining
-        essentially unchanged. It was popularised in the 1960s with the release
-        of Letraset sheets containing Lorem Ipsum passages, and more recently
-        with desktop publishing software like Aldus PageMaker including versions
-        of Lorem Ipsum. centuries, but also the leap into electronic
-        typesetting, remaining essentially unchanged. It was popularised in the
-        1960s with the release of Letraset sheets containing Lorem Ipsum
-        passages, and more recently with desktop publishing software like Aldus
-        PageMaker including versions of Lorem Ipsum.
-      </p>
+    <div className="flex justify-center text-black">
+      <Calendar onChange={onChange} value={value} />
     </div>
   );
 }
-
-export default Calendar;
