--- src/components/home/Time.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-function Time() {
-  return (
-    <div className="pb-4 mt-8 md:mt-0">
-      <p className="text-5xl text-center">3:36 pm</p>
-      <p className="text-3xl text-center">Friday, October 10, 2022</p>
-    </div>
-  );
-}
-
-export default Time;
