+++ src/components/shared/Header.js
@@ -158,7 +158,7 @@ export default function Header() {
   return (
     // Web
     <Box sx={{ flexGrow: 1 }}>
-      <AppBar position="static">
+      <AppBar position="static" style={{ background: "#212534" }}>
         <Toolbar>
           <Typography
             variant="h6"
