+++ src/components/shared/Sidebar.js
@@ -8,7 +8,7 @@ import LogoutIcon from "@mui/icons-material/Logout";
 
 function Sidebar() {
   return (
-    <div className="md:h-screen bg-[#1C1F2E] md:bg-[#131315] pt-2 md:pt-0 fixed bottom-0 left-0 md:static flex md:flex-col w-full md:w-96 justify-center md:justify-between items-center border border-t border-l-0 border-r-0 md:border-r border-[#3f445d] gap-x-2 md:gap-x-0">
+    <div className="md:h-screen bg-[#1C1F2E] md:bg-[#131315] pt-2 md:pt-0 fixed bottom-0 left-0 md:static flex md:flex-col w-full md:w-24 justify-center md:justify-between items-center border border-t border-l-0 border-r-0 md:border-r border-[#3f445d] gap-x-2 md:gap-x-0">
       <div className="md:mt-8 cursor-pointer p-2 md:p-4 hover:bg-[#514FE2] rounded bg-[#514FE2]">
         <HomeIcon />
       </div>
